@@ -6,11 +6,12 @@ authors = ["Char Howland <char@indicio.tech>", "Daniel Bluhm <dbluhm@pm.me>"]
 license = "Apache-2.0"
 
 [tool.poetry.dependencies]
-python = "^3.7"
-pytest = "^6.2.5"
-pytest-asyncio = "^0.16.0"
+python = "^3.9"
+pytest = "^7.4.0"
+pytest-asyncio = "^0.21.1"
 acapy-client = {git = "https://github.com/Indicio-tech/acapy-client.git", rev = "0.7.4"}
 httpx = "^0.15.4"
+controller = {git = "https://github.com/Indicio-tech/acapy-minimal-example.git"}
 
 [tool.poetry.dev-dependencies]
 black = "23.3.0"
@@ -19,3 +20,6 @@ flake8 = "5.0.4"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pyright]
+useLibraryCodeForTypes = true
