@@ -12,33 +12,12 @@ import asyncio
 from base64 import urlsafe_b64decode
 import json
 from os import getenv
-from functools import partial
 
-from acapy_client import Client
-from acapy_client.api.connection import (
-    get_connection,
-    receive_invitation,
-    create_invitation,
-)
-from acapy_client.api.mediation import (
-    get_mediation_requests_mediation_id,
-    post_mediation_request_conn_id,
-    put_mediation_mediation_id_default_mediator,
-    get_mediation_requests,
-)
-from acapy_client.models.conn_record import ConnRecord
-from acapy_client.models.create_invitation_request import CreateInvitationRequest
-from acapy_client.models.mediation_create_request import MediationCreateRequest
-from acapy_client.models.receive_invitation_request import ReceiveInvitationRequest
-from acapy_client.models.get_mediation_requests_state import GetMediationRequestsState
-from acapy_client.models.mediation_record import MediationRecord
-from acapy_client.types import Unset
-
+from controller.controller import Controller
+from controller.models import ConnRecord, InvitationResult, MediationRecord
 from httpx import AsyncClient
 import pytest
-
-from . import record_state
-from . import poll_until_condition
+import pytest_asyncio
 
 
 PROXY = getenv("PROXY", "http://proxy:3000")
@@ -57,100 +36,79 @@ async def get_proxy_invite() -> dict:
         return json.loads(urlsafe_b64decode(url.split("c_i=")[1]))
 
 
-async def get_mediator_invite(external_mediator: Client) -> str:
-    invitation = await create_invitation.asyncio(
-        client=external_mediator, json_body=CreateInvitationRequest()
+async def get_mediator_invite(external_mediator: Controller) -> str:
+    invitation = await external_mediator.post(
+        "/connections/create-invitation",
+        params={"auto_accept": "true"},
+        response=InvitationResult,
     )
-    if not invitation:
-        raise RuntimeError("Failed to retrieve invitation from mediator")
-    assert not isinstance(invitation.invitation_url, Unset)
     return invitation.invitation_url
 
 
-async def proxy_receive_mediator_invite(external_mediator: Client, invite: str):
+async def proxy_receive_mediator_invite(external_mediator: Controller, invite: str):
     async with AsyncClient(timeout=60.0) as client:
         r = await client.post(
             f"{PROXY}/receive_mediator_invitation", json={"invitation_url": invite}
         )
         assert not r.is_error
 
-    # Get mediator mediation requests
-    async def _retrieve():
-        requests = await get_mediation_requests.asyncio(
-            client=external_mediator, state=GetMediationRequestsState.GRANTED
-        )
-        assert requests
-        return requests
+    await external_mediator.record_with_values("mediation", state="granted")
 
-    await poll_until_condition(
-        lambda reqs: bool(not isinstance(reqs.results, Unset) and reqs.results),
-        _retrieve,
+
+async def agent_receive_invitation(agent_bob: Controller, invite: dict) -> ConnRecord:
+    conn_record = await agent_bob.post(
+        "/connections/receive-invitation",
+        json=invite,
+        response=ConnRecord,
     )
 
-
-async def agent_receive_invitation(agent_bob: Client, invite: dict) -> ConnRecord:
-    conn_record = await receive_invitation.asyncio(
-        client=agent_bob, json_body=ReceiveInvitationRequest.from_dict(invite)
+    conn_record = await agent_bob.record_with_values(
+        "connections",
+        record_type=ConnRecord,
+        connection_id=conn_record.connection_id,
+        state="active",
     )
-    if not conn_record:
-        raise RuntimeError("Failed to receive invitation on agent")
-
-    async def _retrieve(connection_id: str) -> ConnRecord:
-        retrieved = await get_connection.asyncio(connection_id, client=agent_bob)
-        assert retrieved
-        return retrieved
-
-    await record_state("active", partial(_retrieve, conn_record.connection_id))
     return conn_record
 
 
-async def agent_request_mediation_from_proxy(agent_bob: Client, conn_id: str):
-    mediation_record = await post_mediation_request_conn_id.asyncio(
-        conn_id=conn_id, client=agent_bob, json_body=MediationCreateRequest()
+async def agent_request_mediation_from_proxy(agent_bob: Controller, conn_id: str):
+    await agent_bob.post(
+        f"/mediation/request/{conn_id}",
     )
-    if not mediation_record:
-        raise RuntimeError(f"Failed to request mediation from {conn_id}")
-
-    async def _retrieve(mediation_record: str) -> MediationRecord:
-        retrieved = await get_mediation_requests_mediation_id.asyncio(
-            mediation_record, client=agent_bob
-        )
-        assert retrieved
-        return retrieved
-
-    await record_state("granted", partial(_retrieve, mediation_record.mediation_id))
+    mediation_record = await agent_bob.record_with_values(
+        "mediation",
+        record_type=MediationRecord,
+        state="granted",
+    )
     return mediation_record
 
 
-async def agent_set_default_mediator(agent_bob: Client, mediation_id: str):
-    result = await put_mediation_mediation_id_default_mediator.asyncio(
-        mediation_id, client=agent_bob
+async def agent_set_default_mediator(agent_bob: Controller, mediation_id: str):
+    result = await agent_bob.put(
+        f"/mediation/{mediation_id}/default-mediator",
     )
-    if not result:
-        raise RuntimeError(f"Failed to set default mediator to {mediation_id}")
     return result
 
 
 @pytest.fixture(scope="session")
 def event_loop():
-    return asyncio.get_event_loop()
+    policy = asyncio.get_event_loop_policy()
+    loop = policy.new_event_loop()
+    yield loop
+    loop.close()
 
 
-@pytest.fixture(autouse=True, scope="session")
+@pytest_asyncio.fixture(autouse=True, scope="session")
 async def setup():
-    agent_bob = Client(base_url=AGENT_BOB)
-    external_mediator = Client(base_url=EXTERNAL_MEDIATOR)
-    mediator_invite = await get_mediator_invite(external_mediator)
-    await proxy_receive_mediator_invite(external_mediator, mediator_invite)
-    invite = await get_proxy_invite()
-    conn_record = await agent_receive_invitation(agent_bob, invite)
-
-    assert conn_record
-    assert isinstance(conn_record.connection_id, str)
-    mediation_record = await agent_request_mediation_from_proxy(
-        agent_bob, conn_record.connection_id
-    )
-
-    assert mediation_record
-    assert isinstance(mediation_record.mediation_id, str)
-    await agent_set_default_mediator(agent_bob, mediation_record.mediation_id)
+    async with Controller(AGENT_BOB) as agent_bob, Controller(
+        EXTERNAL_MEDIATOR
+    ) as external_mediator:
+        mediator_invite = await get_mediator_invite(external_mediator)
+        await proxy_receive_mediator_invite(external_mediator, mediator_invite)
+        invite = await get_proxy_invite()
+        conn_record = await agent_receive_invitation(agent_bob, invite)
+        mediation_record = await agent_request_mediation_from_proxy(
+            agent_bob, conn_record.connection_id
+        )
+        assert mediation_record.mediation_id
+        await agent_set_default_mediator(agent_bob, mediation_record.mediation_id)
